+++ src/main/java/kodi/anime_details/AnimeDetails.java
@@ -1,0 +1,4 @@
+package kodi.anime_details;
+
+public class AnimeDetails {
+}
