+++ src/main/java/kodi/anime_mapping/model/Mapping.java
@@ -1,4 +1,4 @@
-package kodi.mapping.model;
+package kodi.anime_mapping.model;
 
 import lombok.Builder;
 import lombok.Singular;
